Data.Derive.JSON

Description

Derive JSON instances.

Unlike Text.JSON.Generics, single constructor types are not handled specially. Every value is encoded as an object with a single field, with the constructor name as key and the values as its contents.

If the constructor is a record, the contents is an Object with the field names as keys. Otherwise, the contents is an array.

Documentation

makeJSON :: Derivation

Derivation for JSON