Data.Derive.LazySet

Description

A pseudo derivation. For each field in the data type, deriving LazySet generates a function like a record updator, but lazy where possible. This is very useful in certain situations to improve laziness properties. A setter is only lazy if that field is present in one constructor.

Documentation

makeLazySet :: Derivation