Data.DeriveTH

Description

The main TH driver module. It is intended that this need be the only module imported by user code; it takes care of all data threading issues such that all one needs to do is:

   data Foo = Foo ; $( derive makeEq ''Foo )

Documentation

derive :: Derivation -> Name -> Q [Dec]

Derive an instance of some class. derive only derives instances for the type of the argument.

derives :: [Derivation] -> [Name] -> Q [Dec]

deriveFromDec :: Derivation -> Dec -> Q [Dec]

Derive an instance of some class. deriveFromDec only derives instances for the type of the argument.

module Data.Derive.All